MoE restructure: K=60 full-cov GMM, float32 training, 225K dataset
- Dataset: unpack x_diff_data (2-setup rows) → 224,763 single-setup rows
- GMM: K=60 full covariance (BIC-selected), fit on 30K subsample for speed
- Bounding boxes: 5D (n, η, σ_y, W, H) replacing 3D
- train_experts.py: float32 + cholesky_jitter=1e-2, CG threshold→6000,
  fast_computations forced off, GPU memory freed after each cluster
- prepare_data.py: --cov-type arg, GMM_FIT_CAP=30K subsampling
- moe_utils.py: DTYPE changed to float32
- dp_config.py: EXACT_THRESHOLD=8000
- moe_core.py: CG settings in inference context
- bo_collect.py: cross-cluster sharing, zero-waste bonus policy
- select_k.py: new standalone BIC/AIC screening tool
- .gitignore: exclude large workspace files (expert .pt, cluster CSVs, data)
- moe_workspace6/: gmm_gate.joblib + boxes.json committed (experts excluded)
- PROJECT_STATUS.txt: updated to reflect new architecture and current state

+++ DataPipeline/moe_utils.py
@@ -16,7 +16,7 @@ from dp_config import INPUT_COLS, OUTPUT_COLS, LOG_INPUTS
 
 
 # ── Device / dtype ─────────────────────────────────────────────────────────────
-DTYPE  = torch.float64
+DTYPE  = torch.float32
 DEVICE = torch.device("cuda" if torch.cuda.is_available() else "cpu")
